Obligatorisk Oppgave 1 initial commit
I created an initial concept for a TwitterStream wrapper class that
handles all interaction with the Twitter Streaming API client.

The TwitterStreaming class asks for a BlockingQueue (which stores
twitter messages) and then use the Twitter Streaming Client (open
source implementation by Twitter) to fetch the tweets for the given
filter (currently Bergen’s geo location).

It is currently implemented as a runnable object to test the twitter
client. The class should however be seen as a wrapper for a producer
object, and a separate consumer object should be implemented (which
should serve as a skeleton for the students to extend).