Working version using Swing and SwingWorker
I have implemented a working prototype version of the TwitterStream
application.

It implements a TwitterStream class extending SwingWorker. It processes
Tweets from the streaming api client in the background, and calls a
method in the controller that sets the tweet in the view.

In the students’ applications the controller can interpret the JSON and
insert it into a more fancy view.